bench(binary): give the group fan-out sweep its own target
CodSpeed flagged `bench_attr_parser` as +369 Ir/iteration, and it was
real — not the Intel→AMD runner swap I first attributed it to. Measured
locally with callgrind, base 8,605.4 Ir/iteration against head 8,906.3.

None of it is library code. The fan-out fixture instantiates the typed-JID
attribute path (`NodeBuilder::attr::<Jid>`), and adding it to that crate
root shifted inlining enough that `create_attr_node`'s builder stopped
pre-sizing its `SmallVec` and grew through an extra realloc instead: the
delta is `Attrs::insert` +261.7, `SmallVec::try_grow` +254.7, `realloc`
+255 — the owned-node path, which `bench_attr_parser` reaches only from its
setup, never from the `unmarshal_ref` + attr-parse body it measures.

Moving the sweep to its own crate root puts `bench_attr_parser` back at
8,605.4 — exactly the base figure — and leaves `binary_benchmark.rs`
byte-identical to main, so no baseline in it can move. A new fixture should
not be able to shift an unrelated benchmark.
