+++ wacore/binary/benches/binary_benchmark.rs
@@ -3,7 +3,6 @@ use flate2::Compression;
 use flate2::write::ZlibEncoder;
 use std::io::Write;
 use wacore_binary::builder::NodeBuilder;
-use wacore_binary::jid::Jid;
 use wacore_binary::marshal::{
     marshal, marshal_auto, marshal_exact, marshal_ref, marshal_ref_auto, marshal_ref_exact,
     marshal_to, unmarshal_ref,
@@ -59,51 +58,6 @@ fn create_fanout_node() -> Node {
         .attr("type", "text")
         .children(vec![
             NodeBuilder::new("participants").children(devices).build(),
-        ])
-        .build()
-}
-
-/// The `<participants>` shape a sender-key distribution puts on the wire: one
-/// `<to jid=…>` per recipient device, each wrapping its own `<enc>`. This is
-/// the only group stanza whose encode cost is proportional to the recipient
-/// count — a steady-state group send distributes to our own companions only
-/// and carries nothing per member (pinned by
-/// `warm_group_stanza_size_tracks_own_devices_not_group_size` in `wacore`),
-/// which is why the width is swept here rather than assumed.
-///
-/// Recipients are a typed [`Jid`] per device, as `build_participant_node`
-/// passes them, and are spread over distinct users with a handful of devices
-/// each, the way a real fanout resolves. Both details decide the encoding:
-/// a typed JID skips the string classifier production never runs here, and a
-/// device id must stay within `u8` to take the `AD_JID` path — a single user
-/// numbered up to 511 would silently encode half the sweep as `JID_PAIR` and
-/// measure two wire shapes at once.
-fn create_skdm_fanout_node(width: usize) -> Node {
-    const DEVICES_PER_USER: usize = 4;
-    let recipients: Vec<Node> = (0..width)
-        .map(|i| {
-            let user = 5511999990000u64 + (i / DEVICES_PER_USER) as u64;
-            let device = (i % DEVICES_PER_USER) as u16;
-            NodeBuilder::new("to")
-                .attr("jid", Jid::pn_device(user.to_string(), device))
-                .children(vec![
-                    NodeBuilder::new("enc")
-                        .attr("v", "2")
-                        .attr("type", "msg")
-                        .bytes(vec![0xAB; 128])
-                        .build(),
-                ])
-                .build()
-        })
-        .collect();
-    NodeBuilder::new("message")
-        .attr("to", "120363000000000001@g.us")
-        .attr("id", "3EB0A1B2C3D4E5F60718")
-        .attr("type", "text")
-        .children(vec![
-            NodeBuilder::new("participants")
-                .children(recipients)
-                .build(),
         ])
         .build()
 }
@@ -271,20 +225,6 @@ fn bench_marshal_auto_huge_bytes(bencher: divan::Bencher) {
 fn bench_marshal_auto_many_children(bencher: divan::Bencher) {
     bencher
         .with_inputs(create_many_children_node)
-        .bench_refs(|node| black_box(marshal_auto(black_box(node)).unwrap()));
-}
-
-// Group sender-key distribution, swept across the recipient count reported for
-// real groups. Marshalling is linear in the fan-out width, so this is what a
-// cold group send (or a redistribution after a membership change) pays in the
-// encoder; the steady-state send that follows carries no `<participants>` at
-// all. Keeping both facts measurable is what tells a group-size regression
-// ("the warm stanza grew a per-participant node") apart from a group that is
-// merely redistributing.
-#[divan::bench(args = [8, 32, 128, 512])]
-fn bench_marshal_auto_group_fanout(bencher: divan::Bencher, width: usize) {
-    bencher
-        .with_inputs(|| create_skdm_fanout_node(width))
         .bench_refs(|node| black_box(marshal_auto(black_box(node)).unwrap()));
 }
